test/mc.js split into per-sampler files under test/mc/
