+++ .mocharc.yml
@@ -1,1 +1,2 @@
 timeout: 120000
+recursive: true
